+++ projects/ngx-typesafe-forms/src/lib/default-control-value-accessor.ts
@@ -10,7 +10,7 @@ export abstract class DefaultControlValueAccessor<T> implements ControlValueAcce
   public abstract readonly control: AbstractControl<T | null>;
 
   public writeValue(value: T | null): void {
-    this.control.setValue(value);
+    this.control.setValue(value, { emitEvent: false });
   }
 
   public registerOnChange(onChange: (value: T | null) => void): void {
@@ -20,7 +20,7 @@ export abstract class DefaultControlValueAccessor<T> implements ControlValueAcce
   public registerOnTouched(): void {}
 
   public setDisabledState(disabled: boolean): void {
-    disabled ? this.control.disable() : this.control.enable();
+    disabled ? this.control.disable({ emitEvent: false }) : this.control.enable({ emitEvent: false });
   }
 
   public validate(): ValidationErrors | null {