+++ projects/ngx-typesafe-forms/src/lib/form-control.ts
@@ -8,7 +8,7 @@ import {
   abstractControlDirty$,
   abstractControlDisabled$,
   formControlEnabled$,
-  formControlError$,
+  formControlErrors$,
   formControlInvalid$,
   formControlPristine$,
   formControlStatus$,
@@ -77,8 +77,8 @@ export class FormControl<T> extends AngularFormControl implements AbstractContro
     return formControlValue$(this);
   }
 
-  public get error$(): Observable<ValidationErrors | null> {
-    return formControlError$(this);
+  public get errors$(): Observable<ValidationErrors | null> {
+    return formControlErrors$(this);
   }
 
   public get enabled$(): Observable<boolean> {